feat: transfer user id to handle fn to slack integration
