+++ app/tests/integrations/slack/test_users.py
@@ -59,3 +59,44 @@ def test_get_user_locale_without_user_id():
     client = MagicMock()
     user_id = None
     assert users.get_user_locale(client, user_id) == "en-US"
+
+
+def test_replace_user_id_with_valid_handle():
+    assert (
+        users.replace_user_id_with_handle("@user", "Hello <@U12345>, how are you?")
+        == "Hello @user, how are you?"
+    )
+
+
+def test_replace_user_id_with_no_pattern_in_message():
+    assert (
+        users.replace_user_id_with_handle("@user", "Hello user, how are you?")
+        == "Hello user, how are you?"
+    )
+
+
+def test_replace_user_id_with_empty_handle():
+    assert (
+        users.replace_user_id_with_handle("", "Hello <@U12345>, how are you?") is None
+    )
+
+
+def test_replace_user_id_with_empty_message():
+    assert users.replace_user_id_with_handle("@user", "") is None
+
+
+def test_replace_user_id_with_none_handle():
+    assert (
+        users.replace_user_id_with_handle(None, "Hello <@U12345>, how are you?") is None
+    )
+
+
+def test_replace_user_id_with_none_message():
+    assert users.replace_user_id_with_handle("@user", None) is None
+
+
+def test_replace_multiple_user_ids_in_message():
+    assert (
+        users.replace_user_id_with_handle("@user", "Hi <@U12345>, meet <@U67890>")
+        == "Hi @user, meet @user"
+    )
